Move DownloadLink into separate component file
