+++ src/pages/app.tsx
@@ -11,23 +11,7 @@ import Link from 'next/link'
 import { Layout, Heading, Tabs, Badge, ThreeUpText } from '../components'
 import { DemoVideo } from '../components/AppVideo'
 import { CodeSnippet } from '../components/CodeSnippet'
-import { getEventLogger } from '../hooks/eventLogger'
-
-type DownloadLinkProps = React.ComponentProps<typeof Link> & { downloadName: string }
-const DownloadLink: React.FunctionComponent<DownloadLinkProps> = props => {
-    const { downloadName, ...linkProps } = props
-    const handleOnClick = (): void => {
-        const eventArguments = {
-            downloadSource: 'about',
-            downloadName,
-            downloadLinkUrl: linkProps.href,
-        }
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        getEventLogger().log('DownloadClick', eventArguments, eventArguments)
-    }
-
-    return <Link {...linkProps} onClick={handleOnClick} />
-}
+import { DownloadLink } from '../components/DownloadLink'
 
 const AppPage: FunctionComponent = () => {
     const threeUpTextItems = [
